garden: split up the Gantt utils, folder over-crowded
