+++ gui/src/util/UtilGanttScale/UtilGanttScale.test.js
