+++ gui/src/util/UtilGanttScreen/UtilInteractiveGantt.js
