+++ gui/src/util/UtilGanttScreen/UtilInteractiveGantt.test.js
